% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html_tags.R
\name{b}
\alias{b}
\title{Bold text.}
\usage{
b(text)
}
\arguments{
\item{text}{[\code{character}]\cr Text which can be coerced to type character.}
}
\value{
Text in bold.
}
\description{
Text in bold.
}
\examples{
b(text = "text")

}
