% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/include_file_image.R
\name{base64_to_png}
\alias{base64_to_png}
\title{Write base64 to png.}
\usage{
base64_to_png(images, names, force = FALSE)
}
\arguments{
\item{images}{[\code{list}]\cr List of images as base64.}

\item{names}{[\code{character}]\cr Names of images.}

\item{force}{[\code{flag}]\cr TRUE to force writing. Default is FALSE.}
}
\value{
Nothing returned but png created.
}
\description{
Write base64 to png.
}
\examples{
\dontrun{
base64_to_png(images = read_base64_html(name = read_example("Graphics.html")),
              names = "text.png")
}

}
