% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html_tags.R
\name{br}
\alias{br}
\title{Line break after text.}
\usage{
br(text)
}
\arguments{
\item{text}{[\code{character}]\cr Text which can be coerced to type character.}
}
\value{
Text with line break.
}
\description{
Text with line break.
}
\examples{
br(text = "text")

}
