% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quiz_types.R
\name{cloze_mc_list}
\alias{cloze_mc_list}
\title{Multiple cloze multichoice questions with arbitrary number of same
possible answers as list.}
\usage{
cloze_mc_list(conditions, choices, pos_trues, ordered = FALSE)
}
\arguments{
\item{conditions}{[\code{conditions}]\cr Conditions.}

\item{choices}{[\code{character}]\cr Possible choices.}

\item{pos_trues}{[\code{integerish}]\cr Positions of true answers.}

\item{ordered}{[\code{flag}]\cr TRUE for ordered list, FALSE for unordered list,
default is FALSE.}
}
\description{
Multiple cloze multichoice questions with arbitrary number of same
possible answers as list.
}
\examples{
cloze_mc_list(conditions = c("a", "b", "c"), choices = 1:5,
              pos_trues = c(1, 3, 5))

}
