% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quiz_types.R
\name{cloze_mc_multi}
\alias{cloze_mc_multi}
\title{Single cloze multichoice question with arbitrary number of possible answers.}
\usage{
cloze_mc_multi(choices, pos_true, point = 1)
}
\arguments{
\item{choices}{[\code{character}]\cr Possible choices.}

\item{pos_true}{[\code{number}]\cr Position of true answer(s).}

\item{point}{[\code{number}]\cr Point for correct answer. Default is 1.}
}
\value{
Cloze multichoice question as character.
}
\description{
Single cloze multichoice question with arbitrary number of possible answers.
}
\examples{
cloze_mc_multi(choices = 1:5, pos_true = 3)

}
