% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quiz_types.R
\name{cloze_mc_multi_conditions}
\alias{cloze_mc_multi_conditions}
\title{Multiple cloze multichoice questions with arbitrary number of same
possible answers.}
\usage{
cloze_mc_multi_conditions(conditions, choices, pos_trues, point = 1)
}
\arguments{
\item{conditions}{[\code{conditions}]\cr Conditions.}

\item{choices}{[\code{character}]\cr Possible choices.}

\item{pos_trues}{[\code{integerish}]\cr Positions of true answers.}

\item{point}{[\code{number}]\cr Point for correct answer. Default is 1.}
}
\value{
Cloze multichoice questions.
}
\description{
Multiple cloze multichoice questions with arbitrary number of
same parameters.
}
\examples{
cloze_mc_multi_conditions(conditions = c("a", "b", "c"), choices = 1:5,
                          pos_trues = c(1, 3, 5))

}
