% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quiz_types.R
\name{cloze_NM_partial}
\alias{cloze_NM_partial}
\title{Cloze answer of type NM with default partial points.}
\usage{
cloze_NM_partial(
  answer,
  point = 1,
  wrong_default = 7,
  accepted_error = NULL,
  partial_point = 75,
  paf = 10,
  ...
)
}
\arguments{
\item{answer}{[\code{character}]\cr Correct answer.}

\item{point}{[\code{number}]\cr Point for correct answer. Default is 1.}

\item{wrong_default}{[\code{count}]\cr Length of wrong default answer. Default
is 7.}

\item{accepted_error}{[\code{number}]\cr Accepted error. Default is NULL for
no accepted error.}

\item{partial_point}{[\code{number}]\cr Point for partially correct answer.
Default is 75.}

\item{paf}{[\code{number}]\cr factor multiplied with accepted_error for.}

\item{...}{See parameters of cloze_type.}
}
\value{
Cloze answer as character.
}
\description{
Cloze answer of type NM with default partial points.
}
\examples{
cloze_NM_partial(answer = 123)

}
