% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quiz_types.R
\name{cloze_type}
\alias{cloze_type}
\title{Cloze answer of type NM, SA or SAC.}
\usage{
cloze_type(
  type,
  answer,
  point = 1,
  wrong_default = 7,
  accepted_error = NULL,
  partial_answer = NULL,
  partial_point = NULL,
  partial_accept = NULL
)
}
\arguments{
\item{type}{[\code{choice}]\cr Types NM, SA or SAC.}

\item{answer}{[\code{character}]\cr Correct answer.}

\item{point}{[\code{number}]\cr Point for correct answer. Default is 1.}

\item{wrong_default}{[\code{count}]\cr Length of wrong default answer. Default
is 7.}

\item{accepted_error}{[\code{number}]\cr Accepted error. Default is NULL for
no accepted error.}

\item{partial_answer}{[\code{character}]\cr Partially correct answer.}

\item{partial_point}{[\code{number}]\cr Point for partially correct answer.
Default is NULL for no partial point.}

\item{partial_accept}{[\code{number}]\cr Accepted error for partially correct
answer. Default is NULL for no accepted error for partial answer.}
}
\value{
Cloze answer as character.
}
\description{
Cloze answer of type NM, SA or SAC.
}
\examples{
cloze_type(type = "NM", answer = 123)

}
