% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html_tags.R
\name{col_rgb}
\alias{col_rgb}
\title{Colored text.}
\usage{
col_rgb(text, color = "red")
}
\arguments{
\item{text}{[\code{character}]\cr Text which can be coerced to type character.}

\item{color}{[\code{choice}]\cr Name of colors available with grDevices::colors().
Default color is "red".}
}
\value{
Colored text.
}
\description{
Colored text.
}
\examples{
col_rgb(text = "text")

}
