% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quiz_types.R
\name{ddimageortext}
\alias{ddimageortext}
\title{Question of type ddimageortext.}
\usage{
ddimageortext(
  name,
  text,
  image_name,
  drags,
  drops,
  file = NULL,
  grade = 1,
  penalty = 1,
  hidden = 0
)
}
\arguments{
\item{name}{[\code{character}]\cr Name for ddimageortext.}

\item{text}{[\code{character}]\cr Question text for ddimageortext.}

\item{image_name}{[\code{character}]\cr Name for background image.}

\item{drags}{[\code{character}]\cr Drag items.}

\item{drops}{[\code{character}]\cr Drop zones.}

\item{file}{[\code{character}]\cr Optional file encoded as base64. Default is
NULL for no file.}

\item{grade}{[\code{number}]\cr Grade for the question. Default is 1.}

\item{penalty}{[\code{number}]\cr Penalty for the question. Default is 1.}

\item{hidden}{[\code{choice}]\cr 0 for show and 1 for hidden.

Question of type ddimageortext as character.}
}
\description{
Generate a question of type ddimageortext.
}
\examples{
ddimageortext(name = "ddimageortext", text = "Drag and drop",
              image_name = read_example("text.png"),
              drags = drag(number = 1, text = 1, draggroup = 1),
              drops = drop(number = 1, choice = 1, xleft = 1, ytop = 1))

}
