% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quiz_types.R
\name{drag}
\alias{drag}
\title{Drag item.}
\usage{
drag(number, text, draggroup, infinite = TRUE)
}
\arguments{
\item{number}{[\code{count}]\cr Number of drag item.}

\item{text}{[\code{character}]\cr Text of drag item.}

\item{draggroup}{[\code{count}]\cr Group of drag item.}

\item{infinite}{[\code{flag}]\cr TRUE for infinite replicates of drag item.
FALSE for only one drag item. Default is TRUE.}
}
\value{
Drag item as character.
}
\description{
Drag item.
}
\examples{
drag(number = 1, text = 1, draggroup = 1)

}
