% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quiz_types.R
\name{drop}
\alias{drop}
\title{Drop zone.}
\usage{
drop(number, choice, xleft, ytop)
}
\arguments{
\item{number}{[\code{count}]\cr Number of drop zone.}

\item{choice}{[\code{count}]\cr Number of correct drag item.}

\item{xleft}{[\code{number}]\cr Position of drop zone on x-axis.}

\item{ytop}{[\code{number}]\cr Position of drop zone on y-axis.}
}
\value{
Drop zone as character.
}
\description{
Drop zone.
}
\examples{
drop(number = 1, choice = 1, xleft = 1, ytop = 1)

}
