% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quiz_types.R
\name{essay}
\alias{essay}
\title{Question of type essay.}
\usage{
essay(
  name,
  text,
  file = NULL,
  grade = 0,
  penalty = 1,
  hidden = 0,
  responseformat = "editor",
  responserequired = 0,
  responsefieldlines = 5,
  minwordlimit = "",
  maxwordlimit = "",
  attachments = 1,
  attachmentsrequired = 1,
  maxbytes = 0,
  filetypeslist = ".R"
)
}
\arguments{
\item{name}{[\code{character}]\cr Name for essay.}

\item{text}{[\code{character}]\cr Question text for essay.}

\item{file}{[\code{character}]\cr Optional file encoded as base64. Default is
NULL for no file.}

\item{grade}{[\code{number}]\cr Grade for the question. Default is 1.}

\item{penalty}{[\code{number}]\cr Penalty for the question. Default is 1.}

\item{hidden}{[\code{choice}]\cr 0 for show and 1 for hidden.}

\item{responseformat}{[\code{choice}]\cr "plain" or "editor". Default is
editor.}

\item{responserequired}{[\code{choice}]\cr 0 for no response required and 1
for response required. Default is 0.}

\item{responsefieldlines}{[\code{count}]\cr Number of lines for responsefield.}

\item{minwordlimit}{[\code{character}]\cr Minimum word limit for response. Default
is "" for no limit.}

\item{maxwordlimit}{[\code{character}]\cr Maximum word limit for response. Default
is "" for no limit.}

\item{attachments}{[\code{count}]\cr Number of allowed attachments. Default is 1.}

\item{attachmentsrequired}{[\code{count}]\cr Number of required attachments.
Default is 1.}

\item{maxbytes}{[\code{count}]\cr Maximum file size in bytes. Default is 0 for
no limit.}

\item{filetypeslist}{[\code{character}]\cr File types seperated by ", ". Default
is ".R".}
}
\value{
Question of type essay as character.
}
\description{
Generate a question of type essay.
}
\examples{
essay(name = "Essay", text = "Upload your file!")

}
