% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html_tags.R
\name{i}
\alias{i}
\title{Italic text.}
\usage{
i(text)
}
\arguments{
\item{text}{[\code{character}]\cr Text which can be coerced to type character.}
}
\value{
Text in italic.
}
\description{
Text in italic.
}
\examples{
i(text = "text")

}
