% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/include_file_image.R
\name{insert_base64}
\alias{insert_base64}
\title{Insert base64.}
\usage{
insert_base64(name)
}
\arguments{
\item{name}{[\code{character}]\cr Name of file which should be inserted.}
}
\value{
File in base64 encoding.
}
\description{
Insert a file in base64 encoding.
}
\examples{
insert_base64(name = read_example("text.png"))

}
