% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/include_file_image.R
\name{insert_png}
\alias{insert_png}
\title{Insert png in text.}
\usage{
insert_png(name, width = NULL, height = NULL)
}
\arguments{
\item{name}{[\code{character}]\cr Name of file which should be inserted.}

\item{width}{[\code{number}]\cr Width of png. NULL to use original width.
Default is NULL.}

\item{height}{[\code{number}]\cr Height of png. NULL to use original height.
Default is NULL.}
}
\value{
png inserted in text.
}
\description{
Insert png in text.
}
\examples{
insert_png(name = read_example("text.png"))

}
