% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html_tags.R
\name{li}
\alias{li}
\title{Text as list item.}
\usage{
li(text)
}
\arguments{
\item{text}{[\code{character}]\cr Text which can be coerced to type character.}
}
\value{
Text as list item.
}
\description{
Text as list item.
}
\examples{
li(text = "text")

}
