% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html_tags.R
\name{link_file}
\alias{link_file}
\title{Link a file.}
\usage{
link_file(name, text)
}
\arguments{
\item{name}{[\code{character}]\cr Name of file which should be linked.}

\item{text}{[\code{character}]\cr Text which is used as label for the link.}
}
\value{
A link to a file.
}
\description{
Get a link to a file.
}
\examples{
link_file(name = read_example("text.txt"), text = "text.txt")

}
