% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quiz_types.R
\name{multichoice}
\alias{multichoice}
\title{Question of type multichoice.}
\usage{
multichoice(
  name,
  text,
  single,
  answers,
  fractions,
  shuffleanswers = FALSE,
  answernumbering = "none",
  showstandardinstructions = 1,
  file = NULL,
  grade = 1,
  penalty = 1,
  hidden = 0
)
}
\arguments{
\item{name}{[\code{character}]\cr Name for multichoice.}

\item{text}{[\code{character}]\cr Question text for multichoice.}

\item{single}{[\code{flag}]\cr TRUE for single choice. FALSE for multichoice.}

\item{answers}{[\code{character}]\cr Vector of answers.}

\item{fractions}{[\code{character}]\cr Vector of fractions corresponding to
answers.}

\item{shuffleanswers}{[\code{flag}]\cr TRUE to shuffle answers. Default is FALSE.}

\item{answernumbering}{[\code{choice}]\cr "none", "abc" or "123" for numbering.
Default is "none".}

\item{showstandardinstructions}{[\code{choice}]\cr 0 to hide instructions
and 1 to show instructions. Default is 1.}

\item{file}{[\code{character}]\cr Optional file encoded as base64. Default is
NULL for no file.}

\item{grade}{[\code{number}]\cr Grade for the question. Default is 1.}

\item{penalty}{[\code{number}]\cr Penalty for the question. Default is 1.}

\item{hidden}{[\code{choice}]\cr 0 for show and 1 for hidden.}
}
\value{
Question of type multichoice as character.
}
\description{
Generate a question of type multichoice.
}
\examples{
multichoice(name = "Multichoice", text = "Choose a", single = TRUE,
            answers = c("a", "b"), fractions = zero_100_pos(n = 2, pos = 1))

}
