% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html_tags.R
\name{oul}
\alias{oul}
\title{Ordered or unordered list.}
\usage{
oul(items, ordered = FALSE)
}
\arguments{
\item{items}{[\code{character}]\cr Items which can be coerced to type character.}

\item{ordered}{[\code{flag}]\cr TRUE for ordered list. FALSE for unordered list.
Default is FALSE.}
}
\value{
Ordered or unordered list.
}
\description{
Ordered or unordered list.
}
\examples{
oul(items = 1:5)

}
