% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html_tags.R
\name{p}
\alias{p}
\title{Paragraph of text.}
\usage{
p(text)
}
\arguments{
\item{text}{[\code{character}]\cr Text which can be coerced to type character.}
}
\value{
The text in a paragraph.
}
\description{
Put text in paragraph.
}
\examples{
p(text = "text")

}
