% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html_tags.R
\name{p_left}
\alias{p_left}
\title{Left aligned paragraph of text.}
\usage{
p_left(text)
}
\arguments{
\item{text}{[\code{character}]\cr Text which can be coerced to type character.}
}
\value{
Left aligned text.
}
\description{
Left align the text in a paragraph.
}
\examples{
p_left(text = "text")

}
