% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_table.R
\name{parse_table}
\alias{parse_table}
\title{Parse matrix as table.}
\usage{
parse_table(
  m,
  rows = NULL,
  cols = NULL,
  cname = NULL,
  rname = NULL,
  is_graded = TRUE,
  accepted_error = 0,
  wrong_default = 7,
  point = NULL,
  type = "NM",
  partial_answer = NULL,
  partial_point = NULL,
  partial_accept = NULL,
  caption = "",
  width = 100
)
}
\arguments{
\item{m}{[\code{matrix}]\cr Matrix which should be parsed.}

\item{rows}{[\code{count}]\cr Number of rows for output table. Default is NULL
for same number of rows as in m.}

\item{cols}{[\code{count}]\cr Number of columns for output table. Default is NULL
for same number of columns as in m.}

\item{cname}{[\code{character}]\cr Column names.}

\item{rname}{[\code{character}]\cr Row names.}

\item{is_graded}{[\code{flag}]\cr TRUE for grading. FALSE for no grading.}

\item{accepted_error}{[\code{number}]\cr Accepted error. Default is NULL for
no accepted error.}

\item{wrong_default}{[\code{count}]\cr Length of wrong default answer. Default
is 7.}

\item{point}{[\code{number}]\cr Point for correct answer. Default is 1.}

\item{type}{[\code{choice}]\cr Types NM, SA or SAC.}

\item{partial_answer}{[\code{character}]\cr Partially correct answer.}

\item{partial_point}{[\code{number}]\cr Point for partially correct answer.
Default is NULL for no partial point.}

\item{partial_accept}{[\code{number}]\cr Accepted error for partially correct
answer. Default is NULL for no accepted error for partial answer.}

\item{caption}{[\code{character}]\cr Table caption. Default is "" for no caption.}

\item{width}{[\code{number}]\cr Width of table in percent. Default is 100.}
}
\value{
Matrix as character.
}
\description{
Parse a matrix as a string.
}
\examples{
parse_table(m = diag(1))

}
