% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_table.R
\name{parse_table_NM_partial}
\alias{parse_table_NM_partial}
\title{Parse matrix as table with default partial points.}
\usage{
parse_table_NM_partial(
  m,
  accepted_error = 0,
  partial_point = 75,
  paf = 10,
  ...
)
}
\arguments{
\item{m}{[\code{matrix}]\cr Matrix which should be parsed.}

\item{accepted_error}{[\code{number}]\cr Accepted error. Default is NULL for
no accepted error.}

\item{partial_point}{[\code{number}]\cr Point for partially correct answer.
Default is NULL for no partial point.}

\item{paf}{[\code{number}]\cr factor multiplied with accepted_error for.}

\item{...}{See parameters of parse_table.}
}
\value{
Matrix as character with default partial points.
}
\description{
Parse matrix as table with default partial points.
}
\examples{
parse_table_NM_partial(m = diag(1), accepted_error = 0.01)

}
