% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_latex.R
\name{pd}
\alias{pd}
\title{Paste double-dollar around text.}
\usage{
pd(text)
}
\arguments{
\item{text}{[\code{character}]\cr Text which can be coerced to type character.}
}
\value{
The text with double-dollar.
}
\description{
Paste double-dollar around text.
}
\examples{
pd(text = 1:10)

}
