% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_latex.R
\name{pmatrix}
\alias{pmatrix}
\title{Format vector or matrix as LaTeX pmatrix.}
\usage{
pmatrix(values)
}
\arguments{
\item{values}{[\code{numeric}]\cr Vector or matrix of numeric values.}
}
\value{
LaTeX Code for vector or matrix as pmatrix.
}
\description{
Format vector or matrix as LaTeX pmatrix.
}
\examples{
pmatrix(values = c(1, 2))
pmatrix(values = diag(2))

}
