% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/include_file_image.R
\name{read_base64_html}
\alias{read_base64_html}
\title{Read base64 from html file.}
\usage{
read_base64_html(name)
}
\arguments{
\item{name}{[\code{character}]\cr Name of html file which should be read.}
}
\value{
base64 as list.
}
\description{
Read all base64 from html.
}
\examples{
read_base64_html(name = read_example("Graphics.html"))

}
