% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.R
\name{read_example}
\alias{read_example}
\title{Path to example files in extdata.}
\usage{
read_example(name = NULL)
}
\arguments{
\item{name}{[\code{character}]\cr Name of example file. For NULL the names of
all available example files are listed.}
}
\value{
Path to example file(s).
}
\description{
Get the path to files in extdata. This is necessary for the
examples.
}
\examples{
read_example(name = "text.png")

}
