% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_latex.R
\name{read_latex}
\alias{read_latex}
\title{Read LaTeX formulas from arbitrary file with 1 formula per row.}
\usage{
read_latex(file, add_p = TRUE, collapse = FALSE)
}
\arguments{
\item{file}{[\code{character}]\cr File name.}

\item{add_p}{[\code{flag}]\cr TRUE for adding a paragraph per row. Default is TRUE.}

\item{collapse}{[\code{flag}]\cr TRUE to collapse text. Default is FALSE.}
}
\value{
LaTeX formulas as character.
}
\description{
Read LaTeX formulas from a file with 1 formula per row.
}
\details{
At the moment only simple formulas can be read correctly. It may
be necessary to remove some special characters or other LaTeX code before
reading the file.
}
\examples{
read_latex(file = read_example("text.txt"))

}
