% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/include_file_image.R
\name{render_tikz}
\alias{render_tikz}
\title{Render tikz in Rmd to png.}
\usage{
render_tikz(Rmd = "Graphics.Rmd", names = "graphics.png", force = FALSE)
}
\arguments{
\item{Rmd}{[\code{character}]\cr Path to Rmd file with tikz code. See the
example for more information.}

\item{names}{[\code{character}]\cr Names of images.}

\item{force}{[\code{flag}]\cr TRUE to force writing. Default is FALSE.}
}
\value{
Nothing returned but png created.
}
\description{
Render tikz chunks in Rmd-file to png.
}
\details{
It might be necessary to install further R packages and magick to
be able to use this function.

Looking at the file Graphics.Rmd might be useful.
}
\examples{
\dontrun{
render_tikz(Rmd = read_example(name = "Graphics.Rmd"),
            names = c("text1.png", "text2.png"), force = FALSE)
}

}
