% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_html.R
\name{show_html}
\alias{show_html}
\title{Show text as html.}
\usage{
show_html(text)
}
\arguments{
\item{text}{[\code{character}]\cr Text which should be rendered as html.}
}
\value{
Nothing returned but html preview opened.
}
\description{
Show text as html in browser.
}
\details{
This preview is only able to display basic html elements. The
possibility to display more types of Moodle objects would be appreciated.
}
\examples{
\dontrun{
show_html(text = "text")
}

}
