% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html_tags.R
\name{tag}
\alias{tag}
\title{Tagged text.}
\usage{
tag(text, tag)
}
\arguments{
\item{text}{[\code{character}]\cr Text which can be coerced to type character.}

\item{tag}{[\code{character}]\cr Tag name.}
}
\value{
Tagged text.
}
\description{
Add a html-tag to the given text.
}
\examples{
tag(text = "text", tag = "p")

}
