% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_latex.R
\name{write_matrix_latex}
\alias{write_matrix_latex}
\title{Write a matrix to LaTeX.}
\usage{
write_matrix_latex(
  m,
  cname = NULL,
  rname = NULL,
  name = "text.txt",
  force = FALSE,
  append = FALSE,
  write_file = TRUE
)
}
\arguments{
\item{m}{[\code{matrix|vector}]\cr matrix or vector to write to LaTeX.}

\item{cname}{[\code{character}]\cr column name for the matrix. Default is NULL
for no column name.}

\item{rname}{[\code{character}]\cr row name for the matrix. Default is NULL
for no row name.}

\item{name}{[\code{character}]\cr Filename for written file. Default is
"text.txt".}

\item{force}{[\code{flag}]\cr Flag for forcing writing of file. Default is
FALSE.}

\item{append}{[\code{flag}]\cr Flag for appending matrix to file. Default is
FALSE.}

\item{write_file}{[\code{flag}]\cr Flag for writing file. Default is TRUE.
Otherwise the matrix is returned.}
}
\value{
Write matrix as LaTeX.
}
\description{
Write a matrix to LaTeX.
}
\details{
Write a matrix as LaTeX. Further options for formatting the matrix
would be beneficial.
}
\examples{
\dontrun{
  write_matrix_latex(m = diag(2), cname = c("A", "B"), rname = c("A", "B"))
}

}
