% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_moodle_xml.R
\name{write_question_xml}
\alias{write_question_xml}
\title{Write single Moodle question to XML.}
\usage{
write_question_xml(question, name, force = FALSE)
}
\arguments{
\item{question}{[\code{character}]\cr Question as character.}

\item{name}{[\code{character}]\cr Name of XML-file.}

\item{force}{[\code{flag}]\cr TRUE to force writing. Default is FALSE.}
}
\value{
Nothing is returned but an XML-file is written.
}
\description{
Write a single Moodle question to an XML-file.
}
\examples{
\dontrun{
write_question_xml(question = description(name = "Description", text = "text"),
                   name = "description.xml")
}

}
