% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_moodle_xml.R
\name{write_quiz_xml}
\alias{write_quiz_xml}
\title{Write quiz to XML.}
\usage{
write_quiz_xml(name, category, questions, force = FALSE)
}
\arguments{
\item{name}{[\code{character}]\cr Name of XML-file.}

\item{category}{[\code{character}]\cr Name of Moodle category.}

\item{questions}{[\code{character}]\cr Questions as character.}

\item{force}{[\code{flag}]\cr TRUE to force writing. Default is FALSE.}
}
\value{
Nothing is returned but an XML-file is written.
}
\description{
Write a whole quiz to an XML-file.
}
\examples{
\dontrun{
write_quiz_xml(name = "Quiz.xml", category = "Quiz 1",
               questions = description(name = "Description", text = "text"))
}

}
