% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quiz_types.R
\name{zero_100_pos}
\alias{zero_100_pos}
\title{Vector with one 100 and 0 otherwise.}
\usage{
zero_100_pos(n, pos)
}
\arguments{
\item{n}{[\code{count}]\cr Total number of values.}

\item{pos}{[\code{count}]\cr Position of 100.}
}
\value{
Vector with one 100 at position pos and 0 otherwise.
}
\description{
Vector with one 100 at position pos and 0 otherwise.
}
\examples{
zero_100_pos(n = 5, pos = 3)

}
